requirements.txt
google-generativeai
pymupdf
requests
beautifulsoup4
pyyaml
weasyprint
fastapi
uvicorn
python-multipart
python-dotenv